chore: e2e - move style to other file
