+++ example/src/StringVariationScreen.tsx
@@ -1,4 +1,4 @@
-import { View, Text, StyleSheet, ScrollView, SafeAreaView } from 'react-native';
+import { View, Text, ScrollView, SafeAreaView } from 'react-native';
 import {
   useStringVariation,
   useStringVariationDetails,
@@ -6,6 +6,7 @@ import {
 } from '@bucketeer/react-native-client-sdk';
 import { FEATURE_ID_STRING } from '../../e2e/constants';
 import { DetailRow } from './DetailRow';
+import { styles } from './styles';
 
 export default function StringVariationScreen() {
   const { client } = useBucketeerClient();
@@ -76,82 +77,3 @@ export default function StringVariationScreen() {
     </SafeAreaView>
   );
 }
-
-const styles = StyleSheet.create({
-  safeArea: {
-    flex: 1,
-    width: '100%',
-    backgroundColor: '#f8f9fa',
-  },
-  container: {
-    flex: 1,
-    width: '100%',
-    backgroundColor: '#f8f9fa',
-  },
-  header: {
-    backgroundColor: '#fff',
-    padding: 20,
-    borderBottomWidth: 1,
-    borderBottomColor: '#e1e5e9',
-    alignItems: 'center',
-  },
-  readyText: {
-    fontSize: 18,
-    fontWeight: 'bold',
-    color: '#2c3e50',
-  },
-  variationSection: {
-    marginVertical: 8,
-    marginHorizontal: 16,
-    backgroundColor: '#fff',
-    borderRadius: 12,
-    padding: 16,
-    shadowColor: '#000',
-    shadowOffset: {
-      width: 0,
-      height: 2,
-    },
-    shadowOpacity: 0.1,
-    shadowRadius: 3.84,
-    elevation: 5,
-  },
-  variationTitle: {
-    fontSize: 20,
-    fontWeight: 'bold',
-    color: '#2c3e50',
-    marginBottom: 16,
-    borderBottomWidth: 2,
-    borderBottomColor: '#3498db',
-    paddingBottom: 8,
-  },
-  valueContainer: {
-    marginBottom: 16,
-    padding: 12,
-    backgroundColor: '#f8f9fa',
-    borderRadius: 8,
-    borderLeftWidth: 4,
-    borderLeftColor: '#3498db',
-  },
-  valueLabel: {
-    fontSize: 14,
-    fontWeight: '600',
-    color: '#7f8c8d',
-    marginBottom: 4,
-  },
-  value: {
-    fontSize: 16,
-    fontWeight: 'bold',
-    color: '#2c3e50',
-  },
-  detailsContainer: {
-    backgroundColor: '#ecf0f1',
-    borderRadius: 8,
-    padding: 12,
-  },
-  detailsTitle: {
-    fontSize: 16,
-    fontWeight: 'bold',
-    color: '#2c3e50',
-    marginBottom: 12,
-  },
-});
