+++ example/src/styles.tsx
@@ -1,0 +1,108 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  safeArea: {
+    flex: 1,
+    width: '100%',
+    backgroundColor: '#f8f9fa',
+  },
+  container: {
+    flex: 1,
+    width: '100%',
+    backgroundColor: '#f8f9fa',
+  },
+  header: {
+    backgroundColor: '#fff',
+    padding: 20,
+    borderBottomWidth: 1,
+    borderBottomColor: '#e1e5e9',
+    alignItems: 'center',
+  },
+  readyText: {
+    fontSize: 18,
+    fontWeight: 'bold',
+    color: '#2c3e50',
+  },
+  variationSection: {
+    marginVertical: 8,
+    marginHorizontal: 16,
+    backgroundColor: '#fff',
+    borderRadius: 12,
+    padding: 16,
+    shadowColor: '#000',
+    shadowOffset: {
+      width: 0,
+      height: 2,
+    },
+    shadowOpacity: 0.1,
+    shadowRadius: 3.84,
+    elevation: 5,
+  },
+  variationTitle: {
+    fontSize: 20,
+    fontWeight: 'bold',
+    color: '#2c3e50',
+    marginBottom: 16,
+    borderBottomWidth: 2,
+    paddingBottom: 8,
+  },
+  valueContainer: {
+    marginBottom: 16,
+    padding: 12,
+    backgroundColor: '#f8f9fa',
+    borderRadius: 8,
+    borderLeftWidth: 4,
+  },
+  valueLabel: {
+    fontSize: 14,
+    fontWeight: '600',
+    color: '#7f8c8d',
+    marginBottom: 4,
+  },
+  value: {
+    fontSize: 16,
+    fontWeight: 'bold',
+    color: '#2c3e50',
+  },
+  objectDetail: {
+    fontSize: 14,
+    color: '#7f8c8d',
+    marginTop: 4,
+    fontStyle: 'italic',
+  },
+  detailsContainer: {
+    backgroundColor: '#ecf0f1',
+    borderRadius: 8,
+    padding: 12,
+  },
+  detailsTitle: {
+    fontSize: 16,
+    fontWeight: 'bold',
+    color: '#2c3e50',
+    marginBottom: 12,
+  },
+  booleanValueContainer: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    flexWrap: 'wrap',
+  },
+  booleanFeature: {
+    marginLeft: 12,
+    padding: 8,
+    backgroundColor: '#2ecc71',
+    borderRadius: 6,
+  },
+  booleanFeatureText: {
+    color: '#fff',
+    fontSize: 12,
+    fontWeight: 'bold',
+  },
+  booleanDisabled: {
+    marginLeft: 12,
+    color: '#e74c3c',
+    fontSize: 14,
+    fontStyle: 'italic',
+  },
+});
+
+export { styles };
